Added additional constructor to CustomCrafting.
Since I want to shade this plugin and load it from within my plugin, I need to be able to create an instance of the main class while still being able to set the PluginClassLoader, PluginDescription and so on. You might not have a use for this constructor but please leave it in. It helps people, who want to shade the plugin to use it as a library.