Fix 'collect to cursor' action in ECT & Cauldron
Because Spigot has no API to get the affected slots of the action we need to figure it out manually.
In order to keep track of the slots that need to be cleared and updated it uses a weak map, to collect the stack amounts.

This will require an update for WolfyUtils too to work properly, as currently WU executes all the Item Input Buttons instead of just the affected ones.